Partially implement complation for GROUP BY
We now look through the specified groups and determine whether they are an alias
group or a field group, and also handle the non-GROUP-BY case as a special case
in that code.

However, we don't yet distinguish between aggregate and non-aggregate functions
when compiling select fields, so a lot of valid use cases will currently cause
compilation problems.