Change table compilation to include a type context
Instead of just resolving the name of the table to query, we now have a table
expression AST type, and the "type" of a table expression is actually a type
context, which includes both all of the column types and the valid aliases. This
should hopefully be general enough to handle joins, subqueries, and other table
expressions.