Change functions to always take and return lists
New plan for dealing with nested aggregates: aggregates are just regular
functions, and every function maps from list to list. The main thing to handle
with aggregates now is to change the context when evaluating the innermost
aggregate (and non-innermost aggregates are evaluated like regular, although
awkward, functions from a single row to a single row). This approach is also
nice because it lets us selectively transition operations to use numpy in the
future if we need to.