Finish compilation for GROUP BY
We now compile the select fields in two stages. The first stage compiles all
grouped-by fields and gives them access to the table. While we do this, we also
collect the types of those fields to use in the TypeContext for aggregates. We
the compile all other fields using that newly-generated TypeContext and put them
in the right order at the end.