Implement evaluation for subqueries
We no longer evaluate select statements to tables; intead they're always
evaluated to contexts. In the future when the results are written to a table,
that can be done as a separate step. This change involved changing a bunch of
code in tinyquery_test, so I also cleaned up that code to be less repetitive.