Add basic parsing/compiling for table aliases
This needs to be made more sophisticated in the future to account for more
complex dot-separated names, but this seems to work for simple cases.
Dot-separated statements are parsed in a way that it looks like a regular
string with dots in it. When we compile a select with an alias, we replace type
context values to include that alias.