Overhaul the way columns are referenced
Previously, I was doing string operations to separate the (optional) table
component and the column component from a column reference. This doesn't work
for more complicated cases (e.g. figuring out what to do with "a.b.c.d"), so
this commit explicitly uses a table and column name whenever we have information
about a column or reference a column. In many cases, the table is None.

This commit also adds the concept of implicitly-selected columns. For example,
if you select "value + 1" in a subquery, the value column can be referenced in
the outer query, and it looks like the fully-qualified name is allowed. I
implemented this by going through all non-aggregate select fields and adding
them to a special "implicit column context" on the type context. The implicit
columns need to be a special case, since they shouldn't show up in SELECT * and
things like that, and they shouldn't cause ambiguity errors if they conflict
with regular columns. This stuff seems to work pretty well at the compile step,
but the evaluator step probably doesn't handle them correctly. There are also
some known edge cases where we still differ from BigQuery.

I also added some tests that I believe are currently failing. They'll be fixed
in a later commit.