Allow subquery aliases to be used for implicit column references
For now, the approach is to replace all implicitly selected columns with the
alias. In particular, this lets you alias a subquery and use table columns with
the subquery alias (and not the original table alias), although I think there
are a few cases where this still gets things wrong.