Implement basic versions of query and copy jobs
It's now possible to insert query and copy jobs (which runs them synchronously)
and get the status later if we care, and get query results in the format
returned by the BigQuery API.

This also required dot-separated table names for the first time, so I added them
to the parser.

When I was writing this, I realized that tables and contexts have a different
format for column keys (contexts use (table, column) pairs and tables just use
column names), so I fixed up some of the documentation to be consistent with
that.