Add support for IN with a list of literals
I actually just treat it like a regular function, which isn't the most
performant, since it makes a column for each value, but that can be optimized
later if necessary. I still enforce the requirement that they be literals at
parse time, though.

Also fixed a terrifying bug in the equality function for Context that made it so
everything in evaluator_test was artifically returning True. Fortunately, the
only test failures revealed were from typos and things like that.