Implement SELECT *
We parse * into its own special select field, and before compiling select fields
(but after compiling the table expression), we look for any stars and expand
them into uncompiled select fields based on whether we're selecting from a JOIN
or not.

This also fixes a bug in JOIN evaluation where the table component of columns
was being ignored when constructing the data.