Return a simple schema in query results
Summary:
Some of the bq_io code looks at the existence of the schema field to determine
if a query was successful, so this adds an initial version of the schema field
to query results.

No support for nesting or anything like that yet.

Test Plan: tools/runtests.py third_party/tinyquery

Reviewers: colin

Reviewed By: colin

Differential Revision: http://phabricator.khanacademy.org/D12999