Make all tinyquery types nullable
Summary:
Before this change, all booleans and strings were nullable, and all other
types were not nullable. That's inconsistent. This diff changes to make
all types intrinsically nullable, for consistency and also because we'll
be adding a nullable integer type to some of these tests.

Test Plan: run tinyquery tests

Reviewers: benkraft

Subscribers: #logslogslogs

Differential Revision: https://phabricator.khanacademy.org/D28496