Add lexing/parsing support for CASE and CONTAINS
Summary:
I'm working on some query usage analysis where we need to identify fields that
are used in some of our queries.  In order to do this, I'd like to use the
tinyquery parser, but I ran into a query that uses CASE and CONTAINS, neither of
which we have support for in tinyquery.  In the interest of time, I've added
both of these to just the lexer and parser, with a TODO to add compilation (and
evaluation) later.

Test Plan: - `../webapp/tools/runtests.py`

Reviewers: samantha

Reviewed By: samantha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D31335