Implement parsing of multi-way joins
Summary:
Many of our queries that we want to test use multi-way joins, but tinyquery only
supports two-way joins.  This revision is the parser part of an implementation
of multi-way joins.  I set it up so that a join is now a single ast node type
for all kinds of joins, consisting of a single left-hand-side expression (the
FROM part) and a list of one or more right hand side expressions, called
"partial joins" (the JOIN table ON condition) part.  Each partial join has a
table expression, a type (INNER, LEFT OUTER, or CROSS), and for non-cross joins,
a condition.

Note that this revision is not intended to ship by itself, as it breaks two-way
joins in the compiler and evaluator, but I've split this out since it's
relatively self-contained.

Test Plan: - `PYTHONPATH=. ../webapp/tools/runtests.py parser_test`

Reviewers: samantha

Reviewed By: samantha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32037