Implement compilation of multi-way joins
Summary:
In a previous revision, I implemented parsing of joins between more than two
tables.  However, we still lacked the compilation and evaluation stages, and
moreover, we broke the old two-way joins because they too were represented as
mutli-way joins.

This revision implements compilation of multi-way joins.  The implementation was
relatively straightforward, and consisted mainly of changing functions that
acted on two tables to instead act on an arbitrarily-sized array of them.  We
also needed, instead of having a single type parameter for the entire join, to
track a join type for each component in the join, since this could potentially
be different for each.

This is still not yet shippable because we haven't implemented the evaluation
stage, but that will be coming in a separate diff.

Test Plan: `python -m unittest compiler_test`

Reviewers: samantha

Reviewed By: samantha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32211