Implement evaluation of multi-way joins
Summary:
This is the final in the series of diffs implementing multiple-way joins.  The
main change here is that for multi-part joins, I just iterate through the list
of right-hand-sides and then at each step, set the left-hand side to be the
result of the previous join.

I also had to make a couple of changes to the compiler:
1. I needed not to flatten the list of join conditions, and instead have exactly
   one list of (one or more) conditions for that join; otherwise, it was
   difficult to figure out which condition to process with which join.
2. I needed to pass the join type to the function that compiles the join
   conditions, since cross joins are now a unified concept, and those don't have a
   condition (they give [None] for the condition)

I also fixed a bug in the compiler where it could choose the wrong aliased table
if there were two aliases in the same join with the same prefix (which happens
to occur in the evaluator test).

Test Plan: `python -m unittest compiler_test evaluator_test`

Reviewers: samantha

Reviewed By: samantha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32382