Add a script to generate a tinyquery table literal from data in bigquery
Summary:
Creating tinyquery tables by hand is kind of a pain for tables of any
complexity.  This revision adds a script that takes a dataset and table on the
command line and generates a tinyquery table literal (i.e. the python code that
creates a tinyquery.Table object) containing a few rows of data from the
specified table.  This should simplify making test cases for our actual queries.

Note that mode is not yet handled, and record fields will not work either.
Those will be handled later.

Test Plan:
- `tools/make_test_data.py tinyquery test_table`
- `tools/make_test_data.py tinyquery string_table`
- verify that the code generated matches what is in evaluator_test.py

Reviewers: samantha

Reviewed By: samantha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32390