Implement Order By functionality
Summary: The evaluate method for order by uses the sorted method in python.

Also tweaked the code to factor in the possibility of having the list of
order by columns different from the list of selected columns. Precisely,
in case of the number of selected columns being less than the number of
columns in the order by list.

Enabled sorting for strings:
It's nice that python sortings are stable. After playing around with the
python's sort function, I realized that if we reverse the list of
columns based off of which we are sorting and then sort each time
iteratively, it works out pretty well.

Test Plan: python -m unittest discover --pattern=*_test.py

Reviewers: colin, samantha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32568