Implement compilation and evaluation of CASE expressions
Summary:
In D31335, I implemented parsing of CASE expressions, but didn't implement the
ability to evaluate them since for that diff I was primarily interested in using
the parser by itself.  This diff follows up on that and implements evaluating
CASE expressions.  Since CASE expressions can be equivalently expressed by a
series of nested IF calls, and IF calls are already implemented, I chose to have
the compiler convert the CASE to nested IFs and then just compile/evaluate them
using existing code.

One edge case we had to deal with was what happens when none of the clauses in
the CASE evaluates to true, so none of the branches are used.  In this case,
bigquery returns NULL, so I've matched that behavior here.

Test Plan: `python -m unittest compiler_test evaluator_test`

Reviewers: ragini, samantha

Reviewed By: samantha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32419