Implement compilation/evaluation of CONTAINS
Summary:
In D31335, I added parsing support for CONTAINS but didn't implement compilation
or evaluation.  This revision adds compilation and evaluation.  Since CONTAINS
is just implemented as a binary operator, this was fairly straightforward, and I
just needed to add a function to the runtime, which I set up using python's `in`
on strings, which as far as I can tell has the same semantics on strings as
CONTAINS.  (Note in particular that `NULL CONTAINS 'x'` and `'x' CONTAINS NULL`
are both an error in both bigquery and python.)

Test Plan: `python -m unittest evaluator_test`

Reviewers: samantha, ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32426