Implement the bigquery regexp functions
Summary:
Many of our queries use regexp functions, but these weren't implemented yet.
This revision implements the regular expression functions.  To do this, I used
the python `re` module.  This is not the same as the bigquery regexp engine
(which uses the `re2` library), which is roughly a subset of the regexp
functionality in python.  However, I've just added a TODO to use the exact same
engine for now to avoid bringing in the extra dependency that would require
native code.

The bigquery regexp functions are strict about only allowing literals for the
regexp part (and the replacement in REGEXP_REPLACE), so I've added a function
`_ensure_literal` that checks that all values in a column are the same and then
returns that value to match that constraint.

Test Plan: - `python -m unittest evaluator_test`

Reviewers: ragini, samantha

Reviewed By: samantha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32456