Fix a test broken by the merge of the HAVING and multi-way join branches
I added a test for multi-way joins, and in a separate branch, I made it so that
Select expressions now take an extra parameter.  When I landed them both, the
test broke since it had the wrong number of parameters.

This commit adds another `None` parameter to the Select to satisfy the need for
a HAVING parameter.

Test Plan:
`python -m unittest discover --pattern=*_test.py`

Auditors: ragini, samantha