Add a timestamp type and the TIMESTAMP function for creating them.
Summary:
This revision takes the first step in adding all the date and time functions by
creating a timestamp type and TIMESTAMP function.  Bigquery represents
timestamps internally as unix timestamps, but datetimes are easier to work with
in python, so I chose to use datetime.datetime as the internal representation.

`TIMESTAMP` is a bit odd: it can accept either an ISO8601 date/time string, or
an integer corresponding to a unix timestamp expressed in microseconds, instead
of the usual seconds (possibly with decimal places for microseconds).
Amazingly, python doesn't have a function in the standard library for parsing
ISO8601 datetimes (even though it can generate them), so I pulled in the
third-party (pure python) library `arrow` for doing this, which seems to be a
common way to go about parsing them.

Test Plan:
- `pip install -r requirements.txt`
- `python -m unittest discover --pattern=*_test.py`

Reviewers: ragini, samantha

Reviewed By: samantha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32490