Implement extraction and conversion functions for timestamps
Summary:
As part of the effort to fill implement TIMESTAMPs and related functions, this
commit implements all the bigquery functions for extracting various parts of
timestamps or converting them to strings or integers in various ways.

This was a pretty straightforward process of just going down the list of
functions in the bigquery reference and finding how to match the behavior in
python.  The only weird thing was the `WEEK` function, which seemed to have
contradictory definitions in the python and bigquery documentation.  For now I've
left it a TODO to see if there are any weird edge-cases where the behavior
differs.

Test Plan: - `python -m unittest evaluator_test`

Reviewers: samantha, ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32631