Implement the remaining timestamp functions
Summary:
This is the last in a series of commits implementing bigquery's timestamp
functionality.  Here, I've impelemented the remaining miscellaneous functions,
mostly centered around manipulation of unix timestamps.  Since the behavior of
some of these is a little obscure (since they take in and return large integers
with no human-discernable meaning), I took testcases out of the bigquery
documentation to make sure these were working correctly.

The main outstanding question is whether the behavior of python's stftime and
bigquery's version is the same.  It looks like it mostly is (and in turn, they
both appear to take their behavior from C's version of strftime), but it's hard
to be 100% sure.  I've left a TODO to investigate this more later.

Test Plan: `python -m unittest evaluator_test`

Reviewers: samantha, ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32643