Allow several timestamp functions to accept unix or string timestamps
Summary:

In D32631 @samantha pointed out that several of the functions for extracting
parts of a timestamp also work when supplied with integer unix timestamps (with
microseconds) or ISO8601-formatted timestamp strings in bigquery.

This commit adjusts the behavior of these functions in tinyquery to accept any
of these types.  Conveniently, in bigquery, `TIMESTAMP` acts like the identity
function when provided with something that is already a timestamp, so we can
make this change just by composing each of the extraction functions with the
`TIMESTAMP` function.

Test Plan:
`python -m unittest evaluator_test`

Auditors: samantha