Implement LEAST and GREATEST
Summary:
The LEAST and GREATEST functions in bigquery compute the min and max of numbers
across arguments (columns or literals).  This is in contrast to MIN and MAX,
which operate across rows in a group.  This commit implements these functions
using a shared reduce operation.  In contrast to MIN and MAX, which ignore NULL
arguments, LEAST and GREATEST return NULL if any of their arguments are NULL, so
I've added that as a special case.

Test Plan: `python -m unittest evaluator_test`

Reviewers: ragini, samantha

Reviewed By: samantha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32674