Adds null handling to the tinyquery runtime.
Summary:
BigQuery does some fancy things with `NULL` and functions/operators. Previously, we
had simply let TQ crash if it tried to do anything with `None`, but this remedies
most(all?) cases of that.

In addition, there were some subtle bugs that were uncovered as part of null testing,
so this commit also fixes some of those.

On where we differ from BQ, it's perfectly legal to `SELECT TIMESTAMP(NULL)`, but we
crash and burn due to the assigning of `NONETYPE` to the literal. In practice, this
is rare. We don't normally pass literal `NULL` to functions, so it's semi safe to
ignore this functionality for now.

Test Plan: - `python -m unittest evaluator_test`

Reviewers: ragini, colin

Reviewed By: colin

Differential Revision: https://phabricator.khanacademy.org/D32729