Implement record types
Summary:
Bigquery allows record types, which look like potentially nested objects with
properties. This commit implements these in tinyquery.  The implementation
doesn't actually involve nesting, but just represents records as a series of
independent columns (one per leaf field), with a common (dot-delimited) prefix
in the name.  This ends up greatly simplifying most things since we don't need
to special case records anywhere in evaluation, since they're already just plain
columns by that time.

One thing that this representation does complicate is the handling of records
that are repeated at a level below the leaves.  Right now these will end up as
non-repeated fields.  For now I've left this as a TODO.  I plan another diff
that will add basic handling for these such that they just become a repeated
field at the leaf level if there is repetition at any nesting level.  That still
won't get all edge cases (for instance, when there's a 3-level nested structure
with repetition at the outer- and innermost level), but should cover the common
cases that we typically encounter.  Probably what we need to do to fix it
properly is add definition and repetition level properties to each value, akin
to the actual bigquery representation of the data.  But for our purposes, we can
get most of the way there without this complication, so I'll leave that a TODO
until we really need it.

There's also another minor TODO of making the tinyquery API return record fields
as nested when fetching the table info (right now they'll appear as individual
columns).  I'll also handle this in a separate diff.

I also did a minor refactor of some of the functions in tinyquery.py that would
read information from a file, so that now they have a version that reads from a
string, and another version that reads from a file that just does the file IO
and then passes it to the string version.  This makes the core functionality
more easily testable, since we don't need to worry about writing temp files with
test data.

Test Plan: `python -m unittest discover --pattern=*_test.py`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32828