Don't use mock.patch for mocking datetime in the current datetime tests
Summary:

When I substated tinyquery into the KA webapp, the current datetime tests
started to fail, as the mocking of the datetime I added to evaluator_test.py
didn't work correctly with webapp's mocking of datetime.

I found that if I just patch it manually, it works just fine, so this commit
adds code for doing that.  Also added a TODO to investigate why it's not working
with `mock.patch`.

Test Plan:
- `python -m unittest evaluator_test`
- substate this commit into webapp and run:
    `tools/runtests.py third_party/tinyquery/evaluator_test.py`

Auditors: ragini