Implement the LEFT function
Summary:

The LEFT function in bigquery takes the leftmost N characters of a string.  Like
python's slicing, if N is longer than the string, it returns the whole string.
(If the string is null, it returns null.)

This commit implements the LEFT function.  The only tricky bit here is that LEFT
is also a keyword (due to LEFT JOIN), so I had to special-case it as an allowed
function name in the parser.

Test Plan:
- `python -m unittest discover --pattern="*_test.py"`

Auditors: ragini