Allow `STRFTIME_UTC_USEC` to take any timestamp representation
Summary:

Most of the bigquery date and time functions take any of a timestamp, ISO8601
string, or integer unix timestamp in microseconds.  STRFTIME_UTC_USEC already
did this, but the type check did not reflect that fact, so it would lead to type
errors in situations where bigquery would accept the arguments.

This commit adds a type set for the valid timestamp-like types, and converts
an existing usage and STRFTIME_UTC_USEC to use this.

Test plan:
`python -m unittest evaluator_test`

Auditors: ragini