Have `_ensure_literal` handle the case where there are no rows in the result.
Summary:

I ran into a query where with our small set of test data, there are no rows in
the result on which a function with a literal argument is being called.  Since
tinyquery represents a literal as a column with all values equal to the literal,
this means that there was just an empty column at the point where we call
`_ensure_literal`, which caused an exception.

However, this should have been totally fine, since we were just going to throw
away this value anyway.  To fix this, I changed it so that `_ensure_literal`
just returns a sentinal value when it has an empty column.  Note that I didn't
want to use `None` here, since a lot of functions know how to handle `None`, and
we want to make sure this value is not used but just thrown away.

Test Plan:
`tools/runtests.py evaluator_test.py`

Auditors: ragini