Modify tools/make_test_data.py to handle record types
Summary:
In D32390, I wrote a script to fetch some real bigquery data and generate code
for a tinyquery table containing that same data.  However, that was before
record types were implemented, so the script didn't handle those.

This commit adds support for record types (generating separate fields for each
leaf field), so that we can create test fixtures that contain records.

Additionally, I've cleaned up some of the spacing of the generated code, so that
it's easier to inline into a test fixtures file without linter errors.

Test Plan:
- run `tools/make_test_data.py 2016_12_31 ProblemLog`, verify I have a literal
  dumped to stdout that contains the record fields and there are no errors
  reported.

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32907