Fix repeated and timestamp fields when importing newline-delimited JSON.
Summary:

The importer for newline-delimited JSON checks that the value correctly matches
the mode (list for repeated, non-list for non-repeated), but when it actually
calls the function to cast to the correct type, it assumes it has a single value
even when the mode is repeated.

This commit fixes this bug by calling `map(cast_fn, value)` instead of
`cast_fn(value)` when the mode is repeated.

I also realized that the import function for the timestamp type was
`datetime.datetime` which requires multiple positional (numeric) arguments.
Instead I changed it to use arrow, to properly convert from the
ISO8601-formatted strings returned by the bigquery API when exporting
timestamps.

Test Plan:
- substate into webapp, verify it's now possible to load the problem log json
  fixture, which contains repeated fields.
- TODO: tests for the largely untested tinyquery module.

Auditors: ragini