Change the tool for fetching test data to use newline-delimited JSON.
Summary:
In D32923, I changed the format of the webapp test fixtures to use a JSON file
for the schema and newline-delimited JSON for the test data.  This makes it
easier to add additional rows of test data as needed compared to the old
column-based format.  This commit updates the tool for fetching test data to use
the new format.

Test Plan:
- `tools/make_test_data.py 2017_01_07 ProblemLog`
- verify it prints a schema and 5 rows of the specified problem log table.

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32968