Add a command line argument for specifying numer of rows to fetch
Summary:
I'm trying to add additional fixture data to some of our tinyquery tests, and
for this it would be useful to be able to fetch a variable number of rows.

This commit adds a command line parameter for passing the number of rows to
fetch to the test data download script.  I also converted the script to use
argparse for reading command line options, since the combinations were getting
more complicated.

Test Plan:
- `tools/make_test_data.py -n 20 2017_01_07 ProblemLog`
- verify it now prints 20 rows of data

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32982