Do a better check for record fields when importing from json
Summary:
Previously I added a check for records when importing from json by looking at
whether the value was a dict.  However, this doesn't work for nullable records,
since None is not a dict.

This commit instead adds a proper check for whether something is a record field
based on the schema.  To do this, I've refactored the import code a bit so that
we flatten the schema into its tinyquery form before doing anything else
(inserting None at the leaf level, even if None occurs earlier in the
hierarchy).  This cuts down on the amount of awkward checking in the nested
schema when we only have partial dot-separated field names.

Test Plan:
- substate into webapp
- in webapp: `tools/runtests.py bigquery_reports`, verify that this no longer
  fails on rows where a record is null
- `python -m unittest tinyquery_test`

Reviewers: ragini, amy

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33015