Add better handling for repeated record fields on JSON import
Summary:
Previously, when importing from JSON, repetition within a record field only
worked correctly if the leaf of the record had mode repeated.  However, any
level of the record can have repetition, so we need to be smarter about
importing from JSON.

This commit changes import so that if any leaf field in a record or any of its
ancestors is repeated, the final field ends up repeated.  To do this, I
implemented a schema flattening function that sets up the modes and values
correctly prior to filling the columns with data.  Within this function, we
treat records recursively, passing a parameter that indicates whether any
ancestor had been repeated.  When extracting the values, there's a little bit of
complicated logic to make sure that we can only end up with [] and not None for
repeated fields where there are no values.

This is almost but not quite a complete version of bigquery's repeated record
handling.  The one part that's still missing is the interaction of this with
scoped aggregation (when using WITHIN RECORD it may matter at what level of the
hierarchy the repetition occurred).  However, this is still a strict improvement
over what we were doing before, and covers most of the cases we have in
practice, where we're aggregating over an entire record.

Test Plan: `python -m unittest tinyquery_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33110