Allow alternate comment syntax
Summary:
In contrast to many other SQL dialects, bigquery allows any of `--`, `#` or `//`
to start a comment.  We use all of these in our queries, depending on the
preferences of the person writing the query.

This commit modifies the lexer to accept any of these as comments.

Test Plan: `python -m unittest lexer_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33111