Allow number literals to be expressed using scientific notation
Summary:
Bigquery allows integers and floats to be expressed in scientific notation like
`1.4E9` (with either a big or small e).  This commit adds this format to the
allowed number literals for both floats and integers.

Since I tripped myself up along the way with the fact that integers are actually
called `NUMBER` and I had a TODO to rename these to `INTEGER`, I did the rename
to integer in this commit as well.

(parsetab changes are auto-generated)

Test Plan: `SHOULD_REBUILD_PARSER=1 python -m unittest lexer_test parser_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33114