Allow string literals to start with an 'r' sigil.
Summary:
Bigquery allows strings to start with an 'r' sigil, like in python where this
designates a raw string.  It's unclear to me whether bigquery actually treats
these differently and if so how, so for now I've just made this a valid syntax
for plain strings and added a TODO to investigate more.

Test Plan: `python -m unittest lexer_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33116