Implement boolean NOT
Summary:
Boolean NOT was not yet implemented.  This commit adds it as a unary boolean
operator.  Note that in bigquery `NOT NULL` is `NULL`, so I've matched that
behavior and added a note that this is intentional.

(parsetab.py changes are autogenerated)

Test Plan: `python -m unittest parser_test evaluator_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33117