Add some rules of operator precedence.
Summary:
I encountered a situation where tinyquery was giving an error on a query
containing something of the form `<bool> OR x CONTAINS y` because it evaluated
`<bool> OR x` first to get a bool, and then got a type error on contains.

This commit tweaks the operator precedence rules to give and/or very low
precedence and contains higher precedence.

(parsetab changes are auto-generated)

Test Plan: `python -m unittest parser_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33123