Implement JSON_EXTRACT[_SCALAR]
Summary:
This revision implements bigquery's functions for extracting data from JSON
strings, which we use fairly frequently for acessing information we've dumped to
the logs as JSON.

The main difficulty here was handling of nulls, since we need to distinguish
between 'null', which is that literal string in bigquery and represents the
value `null` found in some JSON, and NULL, which is that actual bigquery NULL
value.  I ended up creating a separate object constant to represent bigquery
NULL, so that we could distinguish from python's `None` which is what JSON null
becomes after parsing.

Test Plan: `python -m unittest evaluator_test`

Reviewers: samantha, ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D32665