Allow COUNT(DISTINCT x) to work on repeated fields
Summary:
Trying to run COUNT(DISTINCT x) on a repeated field x currently fails with a
"list type is unhashable error" because tinyquery is trying to add the lists of
values for each row to the set of values, rather than first flattening.

This commit changes it so that we flatten repeated fields before counting
distinct items.

Test Plan: `python -m unittest evaluator_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33159