Make repeated fields survive a join
Summary:
We already handle repeated fields correctly when selecting them normally, but
when we are joining another table (even on a different field), repeated fields
have their mode reset to NULLABLE in the context that results from the join.  To
fix, we just need to pull in the mode from the original column.  Thankfully this
was easy to find since there was already a TODO to fix it!

(Note that I think that this does not yet handle joins ON a repeated field, but
we'll handle that when it comes up.)

Test Plan: `python -m unittest evaluator_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33205