Allow functions on multiple repeated fields so long as repetition matches
Summary:
In D33179, I refactored the runtime so that most functions that normally accept
scalars will also work on repeated fields.  In that revision, I made it so that
if a function was called with multiple repeated fields it was an error, which is
a little more conservative than bigquery's behavior.

Bigquery allows functions on multiple repeated fields if the number of times the
field is repeated in each row matches up.  This frequently occurs when querying
multiple leaves from a repeated record field (but is also allowed whenever this
occurs, record or not).

This commit implements bigquery's behavior more closely.  To do this, instead of
assuming that we have only one repeated field, we now allow there to be an
arbitrary number, and only raise an error when the repetition count in each row
does not match up among the multiple fields.  (I also changed the error message
to match bigquery's when the error does occur.)

Test Plan: `python -m unittest evaluator_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33210